moved database functions out of dedicated file and into utilities to reduce number of files.
