+++ covid_model/analysis/20220525_model_simplification/20220602_calibrate_params.py
@@ -11,9 +11,8 @@ from matplotlib import pyplot as plt
 ### Local Imports ###
 from covid_model import CovidModel
 from covid_model.runnable_functions import do_single_fit, do_create_report
-from covid_model.utils import setup, get_filepath_prefix
+from covid_model.utils import setup, get_filepath_prefix, db_engine
 from covid_model.analysis.charts import plot_transmission_control
-from covid_model.db import db_engine
 
 
 def remove_variants(model, y_dict, remove_variants):
