+++ covid_model/analysis/20220606_GovBriefing/20220608_ba45_scenarios_with_drop_in_reporting.py
@@ -11,9 +11,8 @@ from matplotlib import pyplot as plt
 ### Local Imports ###
 from covid_model import CovidModel
 from covid_model.runnable_functions import do_create_report, do_fit_scenarios, do_create_multiple_reports
-from covid_model.utils import setup, get_filepath_prefix
+from covid_model.utils import setup, get_filepath_prefix, db_engine
 from covid_model.analysis.charts import plot_transmission_control
-from covid_model.db import db_engine
 
 
 def main():
