+++ covid_model/analysis/20220609_performance_improvements/20220609_profile_fit.py
@@ -13,9 +13,8 @@ from matplotlib import pyplot as plt
 ### Local Imports ###
 from covid_model import CovidModel
 from covid_model.runnable_functions import do_single_fit, do_create_report
-from covid_model.utils import setup, get_filepath_prefix
+from covid_model.utils import setup, get_filepath_prefix, db_engine
 from covid_model.analysis.charts import plot_transmission_control
-from covid_model.db import db_engine
 
 
 #@profile
