+++ covid_model/needs_update_or_organizing/run_additional_outputs.py
@@ -5,7 +5,7 @@ import json
 import pandas as pd
 import numpy as np
 ### Local Imports ###
-from covid_model.db import db_engine
+from covid_model.utils import db_engine
 from covid_model.data_imports import get_vaccinations_by_county, ExternalVaccWithProjections, ExternalHosps
