+++ covid_model/needs_update_or_organizing/run_mock_immunity.py
@@ -5,9 +5,8 @@ from cycler import cycler
 import numpy as np
 from matplotlib import pyplot as plt, ticker as mtick
 ### Local Imports ###
-from covid_model.db import db_engine
+from covid_model.utils import db_engine
 from covid_model.model import CovidModel
-from covid_model.cli_specs import ModelSpecsArgumentParser
 
 
 def build_default_model(days):
